pyproject.toml
[project]
name = "chat-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "graphviz>=0.20.3",
    "langchain-chroma>=0.1.4",
    "langchain-community>=0.3.14",
    "langchain-huggingface>=0.1.2",
    "langchain-openai>=0.2.14",
    "langchain-text-splitters>=0.3.4",
    "langchain>=0.3.14",
    "openai>=1.59.3",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sentence-transformers>=3.3.1",
    "sentencepiece>=0.2.0",
    "streamlit-mermaid>=0.3.0",
    "streamlit>=1.41.1",
    "unstructured>=0.16.11",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "neovim>=0.3.1",
]
